search-mcp/requirements.txt
azure-search-documents
mcp
fastapi
uvicorn